Move functions in tests/utilities to PULP_CBC_CMDTest
