+++ pulp/tests/test_pulp.py
@@ -9,11 +9,8 @@ from pulp.apis import *
 from pulp import LpVariable, LpProblem, lpSum, LpConstraintVar, LpFractionConstraint
 from pulp import constants as const
 from pulp.tests.bin_packing_problem import create_bin_packing_problem
-from pulp.tests.utilities import (
-    read_command_line_from_log_file,
-    extract_option_from_command_line,
-)
 from pulp.utilities import makeDict
+import re
 import functools
 import unittest
 
@@ -1444,6 +1441,49 @@ class BaseSolverTest:
 class PULP_CBC_CMDTest(BaseSolverTest.PuLPTest):
     solveInst = PULP_CBC_CMD
 
+    @staticmethod
+    def read_command_line_from_log_file(logPath):
+        """
+        Read from log file the command line executed.
+        """
+        with open(logPath) as fp:
+            for row in fp.readlines():
+                if row.startswith("command line "):
+                    return row
+        raise ValueError(f"Unable to find the command line in {logPath}")
+
+    @staticmethod
+    def extract_option_from_command_line(
+        command_line, option, prefix="-", grp_pattern="[a-zA-Z]+"
+    ):
+        """
+        Extract option value from command line string.
+
+        :param command_line: str that we extract the option value from
+        :param option: str representing the option name (e.g., presolve, sec, etc)
+        :param prefix: str (default: '-')
+        :param grp_pattern: str (default: '[a-zA-Z]+') - regex to capture option value
+
+        :return: option value captured (str); otherwise, None
+
+        example:
+
+        >>> cmd = "cbc model.mps -presolve off -timeMode elapsed -branch"
+        >>> PULP_CBC_CMDTest.extract_option_from_command_line(cmd, "presolve")
+        'off'
+
+        >>> cmd = "cbc model.mps -strong 101 -timeMode elapsed -branch"
+        >>> PULP_CBC_CMDTest.extract_option_from_command_line(cmd, "strong", grp_pattern="\d+")
+        '101'
+        """
+        pattern = re.compile(rf"{prefix}{option}\s+({grp_pattern})\s*")
+        m = pattern.search(command_line)
+        if not m:
+            print(f"{option} not found in {command_line}")
+            return None
+        option_value = m.groups()[0]
+        return option_value
+
     def test_presolve_off(self):
         """
         Test if setting presolve=False in PULP_CBC_CMD adds presolve off to the
@@ -1474,8 +1514,10 @@ class PULP_CBC_CMDTest(BaseSolverTest.PuLPTest):
         if not os.path.getsize(logFilename):
             raise PulpError(f"Test failed for solver: {self.solver}")
         # Extract option_value from command line
-        command_line = read_command_line_from_log_file(logFilename)
-        option_value = extract_option_from_command_line(command_line, option="presolve")
+        command_line = PULP_CBC_CMDTest.read_command_line_from_log_file(logFilename)
+        option_value = PULP_CBC_CMDTest.extract_option_from_command_line(
+            command_line, option="presolve"
+        )
         self.assertEqual("off", option_value)
 
     def test_cuts_on(self):
@@ -1508,12 +1550,14 @@ class PULP_CBC_CMDTest(BaseSolverTest.PuLPTest):
         if not os.path.getsize(logFilename):
             raise PulpError(f"Test failed for solver: {self.solver}")
         # Extract option values from command line
-        command_line = read_command_line_from_log_file(logFilename)
-        gomory_value = extract_option_from_command_line(command_line, option="gomory")
-        knapsack_value = extract_option_from_command_line(
+        command_line = PULP_CBC_CMDTest.read_command_line_from_log_file(logFilename)
+        gomory_value = PULP_CBC_CMDTest.extract_option_from_command_line(
+            command_line, option="gomory"
+        )
+        knapsack_value = PULP_CBC_CMDTest.extract_option_from_command_line(
             command_line, option="knapsack", prefix=""
         )
-        probing_value = extract_option_from_command_line(
+        probing_value = PULP_CBC_CMDTest.extract_option_from_command_line(
             command_line, option="probing", prefix=""
         )
         self.assertListEqual(
@@ -1549,8 +1593,10 @@ class PULP_CBC_CMDTest(BaseSolverTest.PuLPTest):
         if not os.path.getsize(logFilename):
             raise PulpError(f"Test failed for solver: {self.solver}")
         # Extract option value from the command line
-        command_line = read_command_line_from_log_file(logFilename)
-        option_value = extract_option_from_command_line(command_line, option="cuts")
+        command_line = PULP_CBC_CMDTest.read_command_line_from_log_file(logFilename)
+        option_value = PULP_CBC_CMDTest.extract_option_from_command_line(
+            command_line, option="cuts"
+        )
         self.assertEqual("off", option_value)
 
     def test_strong(self):
@@ -1582,8 +1628,8 @@ class PULP_CBC_CMDTest(BaseSolverTest.PuLPTest):
         if not os.path.getsize(logFilename):
             raise PulpError(f"Test failed for solver: {self.solver}")
         # Extract option value from command line
-        command_line = read_command_line_from_log_file(logFilename)
-        option_value = extract_option_from_command_line(
+        command_line = PULP_CBC_CMDTest.read_command_line_from_log_file(logFilename)
+        option_value = PULP_CBC_CMDTest.extract_option_from_command_line(
             command_line, option="strong", grp_pattern="\d+"
         )
         self.assertEqual("10", option_value)
